deleted unncecessary cpp files and create visualization for robin const var

+++ 2D/cpp/rhs12.cpp
@@ -17,13 +17,13 @@ namespace dolfin {
       */
       double ra    = sqrt(  (x[0]-y[0])*(x[0]-y[0])  +  (x[1]-y[1])*(x[1]-y[1])  ) + 1E-13;
             
-      double phi = factor * pow(kappa*ra, (double)nu ) * cyl_bessel_k( nu, kappa*ra );
-      values[0]  = phi * kappa * factor * pow( kappa*ra, (double)nu ) * cyl_bessel_k( nu-1, kappa*ra ) * (x[1] - y[1]) / ra;
+      double phi = factor * pow(kappa*ra, nu ) * cyl_bessel_k( nu, kappa*ra );
+      values[0]  = phi * kappa * factor * pow( kappa*ra, nu ) * cyl_bessel_k( nu-1, kappa*ra ) * (x[1] - y[1]) / ra;
     }
   public:
     const Array<double> x;
     double kappa;
-    int nu;
+    double nu;
     double factor;
   };
 }
